refactor: Rename `lib/plugins/aws/deploy/lib/createStack.js`
