+++ lib/plugins/aws/deploy/index.js
@@ -10,7 +10,7 @@ const checkIfBucketExists = require('../lib/check-if-bucket-exists');
 const getCreateChangeSetParams = require('../lib/get-create-change-set-params');
 const getExecuteChangeSetParams = require('../lib/get-execute-change-set-params');
 const waitForChangeSetCreation = require('../lib/wait-for-change-set-creation');
-const createStack = require('./lib/createStack');
+const createStack = require('./lib/create-stack');
 const cleanupS3Bucket = require('./lib/cleanupS3Bucket');
 const uploadArtifacts = require('./lib/uploadArtifacts');
 const validateTemplate = require('./lib/validateTemplate');
